chore(136): extract shared MCP reset fixture to tests/conftest.py
