+++ tests/test_mcp_tools.py
@@ -34,23 +34,6 @@ from sqlprism.core.mcp_tools import (
 )
 
 
-def _reset_debounce_state():
-    """Reset module-level debounce globals between tests."""
-    _mcp_mod._reindex_pending.clear()
-    for handle in _mcp_mod._reindex_timers.values():
-        handle.cancel()
-    _mcp_mod._reindex_timers.clear()
-
-
-@pytest.fixture(autouse=True)
-def _reset_mcp_state():
-    """Reset global MCP state and debounce state between tests."""
-    _mcp_mod._state = None
-    _reset_debounce_state()
-    yield
-    _mcp_mod._state = None
-    _reset_debounce_state()
-
 # ── 5.1: Pydantic validation and field aliases ──
